+++ src/server/scripts/Outland/black_temple/boss_mother_shahraz.cpp
@@ -97,12 +97,12 @@ struct boss_shahrazAI : public ScriptedAI
         CurrentBeam = rand()%4;                                    // 0 - Sinister, 1 - Vile, 2 - Wicked, 3 - Sinful
         BeamCount = 0;
         FatalAttractionTimer = TIMER_FATAL_ATTRACTION_FIRST;
-        FatalAttractionExplodeTimer = -1;
+		FatalAttractionExplodeTimer = uint32(-1);
         ShriekTimer = TIMER_SILENCING_SHRIEK;
         SaberTimer = TIMER_SABER_LASH;
         RandomYellTimer = TIMER_RANDOM_YELL;
         EnrageTimer = TIMER_ENRAGE;
-        CheckPlayersUndermapTimer = -1;
+        CheckPlayersUndermapTimer = uint32(-1);
         TooFarAwayCheckTimer = 2000;
         checkFatalAttractionDistance = false;
 
@@ -289,8 +289,8 @@ struct boss_shahrazAI : public ScriptedAI
             }
 
             checkFatalAttractionDistance = true;
-            if(clear)
-                FatalAttractionExplodeTimer = -1;
+			if (clear)
+				FatalAttractionExplodeTimer = uint32(-1);
             else
                 FatalAttractionExplodeTimer = 1000;
